+++ includes/install.php
@@ -1,5 +1,6 @@
 <?php
 
+use Give\Form\Migrations\MoveOptionsToVisualAppearanceSection;
 use Give\Log\Migrations\MigrateExistingLogs;
 use Give\Revenue\Migrations\AddPastDonationsToRevenueTable;
 
@@ -159,7 +160,7 @@ function give_run_install() {
 			'v241_remove_sale_logs',
 			'v270_store_stripe_account_for_donation',
 			AddPastDonationsToRevenueTable::id(),
-			MigrateExistingLogs::id(),
+			MigrateExistingLogs::id()
 		];
 
 		foreach ( $upgrade_routines as $upgrade ) {