+++ src/Views/Form/Templates/Sequoia/Actions.php
@@ -25,20 +25,18 @@ class Actions {
 		$this->templateOptions = FormTemplateUtils::getOptions();
 
 		// Get decimal numbers option
-		$decimalNumbersOption = isset( $this->templateOptions['payment_amount']['decimals_enabled'] )
-			? $this->templateOptions['payment_amount']['decimals_enabled']
+		$decimalNumbersOption = isset( $this->templateOptions['visual_appearance']['decimals_enabled'] )
+			? $this->templateOptions['visual_appearance']['decimals_enabled']
 			: 'disabled';
 
 		// Set zero number of decimal.
 		if ( 'disabled' === $decimalNumbersOption ) {
-			add_filter( 'give_get_currency_formatting_settings', [ $this, 'setupZeroNumberOfDecimalInCurrencyFormattingSetting' ], 1 );
+			add_filter( 'give_get_currency_formatting_settings', [
+				$this,
+				'setupZeroNumberOfDecimalInCurrencyFormattingSetting'
+			], 1 );
 			add_filter( 'give_get_option_number_decimals', [ $this, 'setupZeroNumberOfDecimal' ], 1 );
 		}
-
-		// Handle personal section html template.
-		add_action( 'wp_ajax_give_cancel_login', [ $this, 'cancelLoginAjaxHanleder' ], 9 );
-		add_action( 'wp_ajax_nopriv_give_cancel_login', [ $this, 'cancelLoginAjaxHanleder' ], 9 );
-		add_action( 'wp_ajax_nopriv_give_checkout_register', [ $this, 'cancelLoginAjaxHanleder' ], 9 );
 
 		// Handle common hooks.
 		add_action( 'give_donation_form', [ $this, 'loadCommonHooks' ], 9, 2 );
@@ -63,18 +61,22 @@ class Actions {
 	 * @return object
 	 */
 	public function setupStripeBaseStyles( $styles ) {
-		$styles = '{
-			"fontFamily": "Montserrat",
-			"color": "#8d8e8e",
-			"fontWeight": 400,
-			"fontSize": "14px",
-			"::placeholder": {
-			  "color": "#8d8e8e"
-			},
-			":-webkit-autofill": {
-			  "color": "#e39f48"
-			}
-		}';
+		$styles = sprintf(
+			'{
+				"fontFamily": "%1$s",
+				"color": "#8d8e8e",
+				"fontWeight": 400,
+				"fontSize": "14px",
+				"::placeholder": {
+				  "color": "#8d8e8e"
+				},
+				":-webkit-autofill": {
+				  "color": "#e39f48"
+				}
+			}',
+			$this->isGoogleFontEnabled() ? 'Montserrat' : 'system-ui',
+		);
+
 		return json_decode( $styles );
 	}
 
@@ -88,7 +90,9 @@ class Actions {
 	 */
 	public function setupStripeFontStyles( $fontStyles ) {
 		return [
-			'cssSrc' => 'https://fonts.googleapis.com/css2?family=Montserrat&display=swap',
+			'cssSrc' => $this->isGoogleFontEnabled() ?
+				'https://fonts.googleapis.com/css2?family=Montserrat&display=swap' :
+				false,
 		];
 	}
 
@@ -113,20 +117,10 @@ class Actions {
 	 * As per design requirement we want to format donation amount with zero decimal whether or not number of decimal admin setting set to zero.
 	 *
 	 * @since 2.7.0
-	 * @return int|array
+	 * @return int
 	 */
 	public function setupZeroNumberOfDecimal() {
 		return 0;
-	}
-
-	/**
-	 * Handle cancel login and checkout register ajax request.
-	 *
-	 * @since 2.7.0
-	 * @return void
-	 */
-	public function cancelLoginAjaxHanleder() {
-		// add_action( 'give_donation_form_before_personal_info', [ $this, 'getIntroductionSectionTextSubSection' ] );
 	}
 
 	/**
@@ -313,7 +307,21 @@ class Actions {
 				$value['checkout_label']
 			);
 		}
+
 		return $gateways;
+	}
+
+	/**
+	 * @unreleased
+	 * @return bool
+	 */
+	private function isGoogleFontEnabled() {
+		$templateOptions = FormTemplateUtils::getOptions();
+
+		// Set defaults
+		$templateOptions['visual_appearance']['google-fonts'] = ! empty( $templateOptions['visual_appearance']['google-fonts'] ) ? $templateOptions['visual_appearance']['google-fonts'] : 'enabled';
+
+		return give_is_setting_enabled( $templateOptions['visual_appearance']['google-fonts'] );
 	}
 
 }